move write puzzles into io module
