+++ packages/puzzle-fetcher/src/fetching.ts
@@ -1,15 +1,10 @@
 import axios from 'axios'
 import pmap from 'p-map'
-import mkdirp from 'mkdirp'
-import path from 'path'
-import yaml from 'js-yaml'
-import { writeFile } from 'fs/promises'
 import {
   readCompressedBoard,
   parseMoveNotation,
   getEnemySide,
   executeMove,
-  serializeBoard,
 } from '@chessy/core'
 import type { Board, Puzzle, Side, History } from '@chessy/core'
 
@@ -38,25 +33,6 @@ export async function fetchBlunder(): Promise<Blunder> {
 
 function sleep(timeInMillis: number): Promise<void> {
   return new Promise(resolve => setTimeout(resolve, timeInMillis))
-}
-
-export async function writePuzzles(puzzles: Puzzle[], targetDir = './'): Promise<void> {
-  await mkdirp(targetDir)
-  let p = Promise.resolve()
-  puzzles.forEach(puzzle => {
-    const serializedYaml = yaml.dump({
-      ...puzzle,
-      startBoard: serializeBoard(puzzle.startBoard)
-    })
-    p = p.then(() => {
-        writeFile(
-        path.join(targetDir, puzzle.id + '.yml'),
-        serializedYaml,
-        'utf-8',
-      )
-    })
-  })
-  return p
 }
 
 export function parseBlunder(blunder: Blunder): Puzzle {
